requirements.txt
asgiref==3.2.7
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
Markdown==3.2.1
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.1
